requirements.txt
pyVmomi>=8.0.0.0
python-decouple>=3.8
